improve get_from_nv stability

+++ gaugan/get_from_nv.py
@@ -22,11 +22,19 @@ class run_request ():
         self.tempfile_name = 'temp_%s.sh' % self.name
         with open(self.tempfile_name, 'w+') as f:
             f.write(self.request)
-        self.p = subprocess.Popen('bash %s' % self.tempfile_name, shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE,encoding="utf-8")
-        print("退出：" + self.name)
+        if self.name.startswith('request1'):
+            self.p = subprocess.Popen('bash %s' % self.tempfile_name, shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE,encoding="utf-8")
+            print("正在运行:" + self.name)
+            # self.p = None
+            # os.system('bash %s' % self.tempfile_name)
+        else:
+            self.p = None
+            os.system('bash %s' % self.tempfile_name)
+            print("退出：" + self.name)
 
     def stop(self):
-        self.p.kill()
+        if self.p is not None:
+            self.p.kill()
         if os.path.exists(self.tempfile_name):
             os.remove(self.tempfile_name)
         print("退出：" + self.name)
@@ -34,7 +42,10 @@ class run_request ():
 
 def generate_result_from_nv(input_file, output_file):
     print('Trying to get from nv')
-    template1 = r'''curl 'http://54.244.111.218:443/nvidia_gaugan_submit_map' \
+    template1 = r'''export http_proxy='127.0.0.1:8889'
+    export https_proxy=$http_proxy
+    echo "Using proxy at $http_proxy"
+    curl 'http://54.244.111.218:443/nvidia_gaugan_submit_map' \
   -H 'Connection: keep-alive' \
   -H 'Accept: */*' \
   -H 'User-Agent: Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4400.0 Safari/537.36 Edg/90.0.782.0' \
@@ -45,7 +56,10 @@ def generate_result_from_nv(input_file, output_file):
   --data-raw '__DATA__' \
   --compressed \
   --insecure'''
-    template2 = r"""curl 'http://54.244.111.218:443/nvidia_gaugan_receive_image' \
+    template2 = r"""export http_proxy='127.0.0.1:8889'
+    export https_proxy=$http_proxy
+    echo "Using proxy at $http_proxy"
+    curl 'http://54.244.111.218:443/nvidia_gaugan_receive_image' \
   -H 'Connection: keep-alive' \
   -H 'User-Agent: Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4400.0 Safari/537.36 Edg/90.0.782.0' \
   -H 'Content-Type: multipart/form-data; boundary=----WebKitFormBoundaryEls912uuOpbLfYzg' \
@@ -67,7 +81,7 @@ def generate_result_from_nv(input_file, output_file):
     match_obj = re.match(r"b'(.*)'", str(img_in_base64))
 
     name = time.strftime('%Y/%m/%d', time.localtime()) + \
-        ',' + str(time.time())
+        ',' + str(time.time()).replace('.','-')
     params = {
         'imageBase64': 'data:image/png;base64,' + match_obj.group(1),
         # 'imageBase64' : 'data:image/png;base64,' + img_in_base64,
@@ -79,13 +93,12 @@ def generate_result_from_nv(input_file, output_file):
     request2 = template2.replace('__NAME__', name).replace(
         '__OUTPUT_FILE__', output_file)
 
-    th_req1 = run_request(request1, 'request1-'+str(time.time()))
-    th_req2 = run_request(request2, 'request2-'+str(time.time()))
+    th_req1 = run_request(request1, 'request1-'+name.split(',')[1])
+    th_req2 = run_request(request2, 'request2-'+name.split(',')[1])
 
     th_req1.run()
-    time.sleep(3.0)
+    time.sleep(4.0)
     th_req2.run()
-    time.sleep(3.0)
 
     th_req1.stop()
     th_req2.stop()
